+++ EarthBender.h
@@ -8,19 +8,19 @@ class EarthBender:public Persona{
 
    private:
         int colesCosechadas;
-        int graduacion;
+        double graduacion;
         PoderEspecial* poderEspecial;
 
    public:
        EarthBender();
-       EarthBender(int colesCosechadas, int graduacion, PoderEspecial* poderEspecial, string nOrigen,string nombreP,int edadP,char s);
+       EarthBender(int colesCosechadas, double graduacion, PoderEspecial* poderEspecial, string nOrigen,string nombreP,int edadP,char s);
         ~EarthBender();
 
         int getcolesCosechadas();
        void setcolesCosechadas(int);
 
-        int getGraduacion();
-       void setGraduacion(int);
+        double getGraduacion();
+       void setGraduacion(double);
 
         PoderEspecial* getPoderEspecial();
        void setPoderEspecial(PoderEspecial*);
